requirements.txt
streamlit==1.29.0
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.18
langchain-anthropic==0.3.14
langchain-google-genai==2.1.5
streamlit-authenticator==0.1.4
pyyaml==6.0.1
streamlit-option-menu==0.3.12
gradio==5.32.0